atualização e correção

+++ Game Server/Maps/MapData.cs
@@ -61,7 +61,7 @@ namespace GameServer.Maps {
             }
         }
 
-        public void SendPlayerMove(PlayerData pData, int dir) {
+        public void SendPlayerMove(PlayerData pData, short dir) {
             foreach (int playerID in PlayerID) {
                 //procura o jogador pelo ID
                 var playerData = Authentication.FindByAccountID(playerID);
